feat: support `OpenaiAssistantAgent`
