+++ .npmrc
@@ -2,3 +2,4 @@ auto-install-peers = true
 enable-pre-post-scripts = true
 prefer-workspace-packages = true
 save-workspace-protocol = true
+link-workspace-packages = true
